refactor: record 디자인 파일 통합
